src buffer read using movdqu: src alignment is not necessary

+++ PlayPcmWin/00experiments/SseCopyTest64/PCM16to24.cpp
@@ -10,7 +10,7 @@ PCM16to24(const int16_t *src, uint8_t *dst, int64_t pcmCount)
     }
 
     // ASM実装は16の倍数サンプル単位の処理。端数をC++で処理します。
-    int countRemainder = pcmCount % 16;
+    int64_t countRemainder = pcmCount % 16;
     int64_t countAsm = pcmCount - countRemainder;
 
     CpuCapability cc;
